ADAHUB-1205: Sharman Bringup [PandM]
Reason for change:Firmware Upgrade support for
SR213 Sharman.

Test Procedure: Build should work
Risks:Low

(cherry picked from commit 3e6934620ea1419f8590e696a8f169b65e562b71)

+++ source/TR-181/middle_layer_src/cosa_deviceinfo_dml.c
@@ -14348,6 +14348,11 @@ Xconf_SetParamBoolValue
                            v_secure_system ("kill -9 `pidof cbr_firmwareDwnld.sh `");
                        }
                            status = v_secure_system("/etc/cbr_firmwareDwnld.sh &");
+#elif defined(_SR213_PRODUCT_REQ_)
+            if(0 == v_secure_system("pidof sr213_firmwareDwnld.sh"))  {
+                           v_secure_system ("kill -9 `pidof sr213_firmwareDwnld.sh `");
+                       }
+                       status = v_secure_system("/etc/sr213_firmwareDwnld.sh &");
 #elif defined(_SE501_PRODUCT_REQ_)
                         if(0 == v_secure_system("pidof se501_firmwareDwnld.sh"))  {
                            v_secure_system ("kill -9 `pidof se501_firmwareDwnld.sh `");
